fix: Cannot quit app when there is a dirty editor (#13164)
Contributed by STMicroelectronics
